modules/cvparser/requirements.txt
tika==1.23
fastapi==0.88.0
uvicorn==0.22.0
numpy==1.22.0
pdfminer==20191020
pdfplumber
seqeval==1.2.2
torch
tqdm==4.46.0
transformers==4.30.0
scikit-learn
opencv-python
poppler-utils
pdf2image
PyYAML==6.0
PyJWT